Create the TemplatedEmitter class.
This class simply substitutes the output with an environment before
sending it upstream to the StructuredEmitter class. This was motivated
by the fact that I found it very helpful to use templates during code
generation and I wanted the substitution to be abstracted away from the
hardcore code generation process. It is always possible to bypass the template
substitution by using the emit_raw() method.