Rewrite the name resolving pass.
This pass takes care of resolving which object are identifiers bound to.
It then stores the result by adding a 'resolved' attribute to each identifier
and qualified identifier node in the tree. These 'resolved' attributes can then
be accessed to refer to the object bound to the identifier, which is
invaluable during code generation.