Reimplement the AST.
Rebuild the AST without using automatic code generation from pycparser.
Although the new implementation is almost a copy of the one found in pycparser,
I find it more flexible. It is now easier to add children and attributes to
a node dynamically.

A NodeTransformer class is also added in order to make it easy to transform the
AST from one pass to another.

Also added the EzNode decorator to create node classes easily.

Removed the (now useless) config files, build file and automatically built AST.

Corrected a bug in the Parser where we passed a reverse iterator as a children
list. The use of an iterator made us consume the iterator the first time we
used it. Now using a plain list.