Restructure the project folder organization.
The folder hierarchy now models (not perfectly!) the hierarchy of Clang.
Since this project is WAY smaller than Clang, this change does not mean
that the Clang way will always prevail.

Right now, most subpackages are empty. It does however give a clear idea of
how things plan to be organized.